+++ tests/Feature/SettingsTest.php
@@ -76,4 +76,37 @@ class SettingsTest extends FeatureTestCase
 
         $response->assertSee('Login');
     }
+
+    public function test_it_updates_the_default_profile_view()
+    {
+        $user = $this->signin();
+
+        $response = $this->json('POST', '/settings/updateDefaultProfileView', [
+            'name' => 'life-events',
+        ]);
+
+        $response->assertStatus(200);
+
+        $this->assertDatabaseHas('users', [
+            'profile_active_tab' => 'life-events',
+            'id' => $user->id,
+        ]);
+
+        $response = $this->json('POST', '/settings/updateDefaultProfileView', [
+            'name' => 'notes',
+        ]);
+
+        $response->assertStatus(200);
+
+        $this->assertDatabaseHas('users', [
+            'profile_active_tab' => 'notes',
+            'id' => $user->id,
+        ]);
+
+        $response = $this->json('POST', '/settings/updateDefaultProfileView', [
+            'name' => 'nawak',
+        ]);
+
+        $response->assertStatus(200);
+    }
 }
